+++ src/Discussion/DiscussionService.php
@@ -12,7 +12,10 @@ use GuzzleHttp\RequestOptions;
 use Wizaplace\SDK\AbstractService;
 use Wizaplace\SDK\Authentication\AuthenticationRequired;
 use Wizaplace\SDK\Catalog\DeclinationId;
+use Wizaplace\SDK\Exception\CompanyHasNoAdministrator;
 use Wizaplace\SDK\Exception\NotFound;
+use Wizaplace\SDK\Exception\ProductNotFound;
+use Wizaplace\SDK\Exception\SenderIsAlsoRecipient;
 use Wizaplace\SDK\Exception\SomeParametersAreInvalid;
 
 /**
@@ -76,26 +79,26 @@ final class DiscussionService extends AbstractService
     /**
      * Start a discussion with a vendor about a specific product.
      *
-     * @throws NotFound
+     * @throws ProductNotFound
+     * @throws SenderIsAlsoRecipient
+     * @throws CompanyHasNoAdministrator
      * @throws AuthenticationRequired
      */
     public function startDiscussion(int $productId): Discussion
     {
         $this->client->mustBeAuthenticated();
 
-        try {
-            $discussionData = $this->client->post('discussions', [RequestOptions::JSON => ['productId' => $productId]]);
+        $discussionData = $this->client->post('discussions', [RequestOptions::JSON => ['productId' => $productId]]);
 
-            return new Discussion($discussionData);
-        } catch (ClientException $e) {
-            throw new NotFound('The product '.$productId.' has not been found.', $e);
-        }
+        return new Discussion($discussionData);
     }
 
     /**
-     * Start a discussion with a vendor about a specific product.
+     * Start a discussion with a vendor.
      *
      * @throws NotFound
+     * @throws SenderIsAlsoRecipient
+     * @throws CompanyHasNoAdministrator
      * @throws AuthenticationRequired
      */
     public function startDiscussionWithVendor(int $companyId): Discussion
@@ -114,7 +117,9 @@ final class DiscussionService extends AbstractService
     /**
      * Start a discussion with a vendor about a specific product identified by its declination ID.
      *
-     * @throws NotFound
+     * @throws ProductNotFound
+     * @throws SenderIsAlsoRecipient
+     * @throws CompanyHasNoAdministrator
      * @throws AuthenticationRequired
      */
     public function startDiscussionFromDeclinationId(DeclinationId $declinationId): Discussion
@@ -125,8 +130,9 @@ final class DiscussionService extends AbstractService
             $discussionData = $this->client->post('discussions', [RequestOptions::JSON => ['declinationId' => (string) $declinationId]]);
 
             return new Discussion($discussionData);
-        } catch (ClientException $e) {
-            throw new NotFound('The product with declination '.$declinationId.' has not been found.', $e);
+        } catch (ProductNotFound $e) {
+            // add declinationId to exception's context
+            throw new ProductNotFound($e->getMessage(), array_merge($e->getContext(), ['declinationId' => (string) $declinationId]), $e);
         }
     }