+++ src/Exception/ErrorCode.php
@@ -14,10 +14,18 @@ use MyCLabs\Enum\Enum;
  * @method static ErrorCode BASKET_NOT_FOUND()
  * @method static ErrorCode COUPON_CODE_DOES_NOT_APPLY()
  * @method static ErrorCode COUPON_CODE_ALREADY_APPLIED()
+ * @method static ErrorCode PRODUCT_NOT_FOUND()
+ * @method static ErrorCode REVIEWS_ARE_DISABLED()
+ * @method static ErrorCode SENDER_IS_ALSO_RECIPIENT()
+ * @method static ErrorCode COMPANY_HAS_NO_ADMINISTRATOR()
  */
 final class ErrorCode extends Enum
 {
     private const BASKET_NOT_FOUND = 1;
     private const COUPON_CODE_DOES_NOT_APPLY = 2;
     private const COUPON_CODE_ALREADY_APPLIED = 3;
+    private const PRODUCT_NOT_FOUND = 4;
+    private const REVIEWS_ARE_DISABLED = 5;
+    private const SENDER_IS_ALSO_RECIPIENT = 6;
+    private const COMPANY_HAS_NO_ADMINISTRATOR = 7;
 }