+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 
 ### New features
 
+ - Introduce a new detailed exception `\Wizaplace\SDK\Exception\ProductNotFound`
+ - Introduce a new detailed exception `\Wizaplace\SDK\Exception\ReviewsAreDisabled`
+ - Introduce a new detailed exception `\Wizaplace\SDK\Exception\SenderIsAlsoRecipient`
+ - Introduce a new detailed exception `\Wizaplace\SDK\Exception\CompanyHasNoAdministrator`
  - Added `\Wizaplace\SDK\Catalog\CatalogServiceInterface` which is implemented by `\Wizaplace\SDK\Catalog\CatalogService`
  - Added `\Wizaplace\SDK\Catalog\AbstractCatalogServiceDecorator` which allows to easily decorate any `\Wizaplace\SDK\Catalog\CatalogServiceInterface`